Bump kured image version
The kured image is bumped because the kured package has introduced
a new required patch to support current 1.16 k8s and avoid deprecated
and deleted API methods. Since this was a patch no version bump
happend in kured package, thus the resolved tag remains the same
unless we bump the image version manually.